Fix so that test passes on all storage formats. ()
The test `creatingSemanticSearchIndexesOnMultipleLabelsMustBePossible`
shouldn't really have passed since it requires kernel version
GLORIOUS_FUTURE and that we enable single-stage filtering. However,
because of a shortcut in record format, in
`IntegrityValidator::validateSchemaRule` it was never validated and the
test passed when using record format, but failed when using block
format.

This PR updates Operations::indexCreate so that we always fail on
invalid configuration and we also update the test to use the required
configurations.