Fixes an issue in EncodingIdMapper for mixed String/long values ()
The issue was that for e.g. --id-type=STRING, but with an :ID column
forcing the actual ID types to be of another type, e.g. long and if
there were accidental hash collisions between values they would get
stored in the "collision values cache" in a different form (in this case
as String), but upon lookup compared to a long value and would therefor
always mismatch. This would result in some node IDs that couldn't be
found later for relationship import.

Also made these collision values typically occupy slightly less space in
memory.