Unflake LuceneDirectoryTest on Windows ()
- Refactor tests to make sure they use the same index config/setup
- Change the merging parameters so that Lucene does not merge during a
`commit`

This is mostly as before, changed the minMergeMB to a value that is
definitely not exceeded by the test documents.

- Check for log messages using the _eventually_ variant:

Since logs are produced during a merge which happens on a different
thread,
those messages were not always immediately visible to the
AssertableLogProvider.
I had this happen locally about once every ~4k test runs and a single
retry of
10ms (which is hardcoded in the _eventually_ methods) was enough to see
the
message. A 500ms deadline seems reasonable to allow for CI to be slow as
CI
does and not wait forever.

Funnily enough, whenever this failed, the message were then visible by
the call to `serialize` that would render to actual message, leading to
errors such as 'did not find log "foo" in these logs: ["foo"]'.

- Wrap the try-with-resources around `indexWriter::close` instead of
`forceMerge`

Every now and then (but for some reason, not always), Lucene would fail
twice,
once during `forceMerge` and again when calling `close`. We only tested
on the
`forceMerge` exception, so moving the `assertThatThrownBy` up to the
`close`
level will catch both instances.

- Skip failing test on Windows

The tests sets POSIX permissions to force the test to fail, which does
not work on Windows or other non-POSIX OSs. On those systems, we bail
out and skip the test.

part of IND-81
part of IND-82
part of IND-83