Index reader API extension/refactor to support vector SSF parameters in queries ()
Re-order `query` / `progressor` / `createQuery` / `newQueryContext()`
method parameters, to support multiple `PropertyIndexQuery…` predicates
as the final (varargs) parameter.

Add a small amount of validation in the vector index.

Contributes to IND-2

This refactor opens up the option of passing multiple extra queries
(`PropertyIndexQuery`) from the kernel API level,
down to the index implementation level, in particular to allow the
enhancement of vector index queries with single stage
filter queries (exact predicates or range predicates).

A later PR will add testing at kernel API of single-state-filter
queries; that change is kept separate so that this one is focused solely
as an enabling refactor which should allow existing indexes (vector and
otherwise) to continue to function as before.