Remove concept of changed properties for nodes and relationships, and don't read old property on set ()
This removes the concept of changed properties in txstate. Now we add
every change from `setProperty` as an added property instead of
splitting it into added and changed in `Operations`. This allowed us to
remove the read of the previous value on every `setProperty`.
A result of this is that we now always add a command on sepProperty,
even when we set it to the same value. And a consequence from this is
that if you have security and not allowed to set a property you are
always blocked, where you before were allowed to set a property as long
as the value didn't change.
Another thing that is affected is when you have `CALL` in transactions
or concurrent transactions with only set property that could have been
no-ops before are now always a change. In `SnapshotExecutionIT` one test
always failed due to this since the page now always was marked as dirty.
And in `MergeLockConcurrencyTest` the transactions will deadlock since
both transactions takes locks where the old way resulted in a no-op not
needing exclusive index/label locks

Spd compat build:
https://live.neo4j-build.io/buildConfiguration/BuildServicePipelines_Neo4jSpdIntegrationTests/26760341

Benchmarks (pokec_write)
https://benchmarking.neo4j.today/index.html?route=_dashboards_branch-form&personal-projects=EmelieBondesson/2025.11-remove-property-read-on-set&public-projects=dev

cl: Every property change will now result in a change for the database
and in the transaction log, even setting the value to the existing one,
resolves CVE-2025-12738

---------

Co-authored-by: Valdemar Roxling <valdemar.roxling@neo4j.com>