Dynamic relationship types for pipelined & parallel ()
Non-fused and fused now share the same base implementation as
interpreted and slotted, in the same way that the algorithm was shared
by dynamic node labels.

This PR also contains a refactor to add a `ClosingRelationshipIterator`,
which is nothing more than an empty abstract base class that can be used
in place of the commonly found `ClosingLongIterator with
RelationshipIterator`, saving approximately 18 bytes of source code
storage per usage, which might even be enough to amortise the cost of
writing it.