MergeUniqueNode ()
For plans like

```
.merge("CREATE (a:L(id=1234))")
.nodeLockingUniqueSeek("a:L(id=1234)")
```

we do a lot of necessary seeking, What we roughly do is

- grab a shared index lock
- seek for a:L(id=1234)
- if we don't find the node, upgrade the lock to an exclusive lock
- seek for a:L(id=1234) again
- if it is not there we create a:L(id=1234) and in the process we also
acquire an exclusive node lock (and a shared label lock)
- When we have created the node we directly do yet another seek to
verify that we didn't violate the uniqueness constraint

The last seek is completely unnecessary since we are holding an
exclusive lock and just verified that no such node existed.

By creating a new operator `MergeUniqueNode` and adding a new method to
kernel API we can minimize the number of seeks we do for each operation.

~Builds on top of ~


[Benchmarks](https://benchmarking.neo4j.today/?route=_dashboards_branch-form&personal-projects=pontusmelke/ind-45-upsert-unique-node&public-projects=dev&deselected=%5B%22statistical-fn%22+%22max%22%5D&deselected=%5B%22statistical-fn%22+%22min%22%5D&deselected=%5B%22statistical-fn%22+%22perc_25%22%5D&deselected=%5B%22statistical-fn%22+%22perc_50%22%5D&deselected=%5B%22statistical-fn%22+%22perc_75%22%5D&deselected=%5B%22statistical-fn%22+%22perc_90%22%5D&deselected=%5B%22statistical-fn%22+%22perc_95%22%5D&deselected=%5B%22statistical-fn%22+%22perc_99%22%5D&deselected=%5B%22statistical-fn%22+%22perc_99_9%22%5D&deselected=%5B%22statistical-fn%22+%22sample_size%22%5D&group-by=branch):
no regressions and some improvements in the ~30% range.

changelog: Add new operator MergeUniqueNode speeding up some merge
queries when there are uniqueness constraints.