Feat(PLAN-2733) - SPD: Cached property boundaries due to mutating patterns invalidating them ()
If a mutating pattern **invalidates** the cached properties, then any
properties beyond this point should *not* be gathered in
`findGlobalPropertyAccessesWithContext`. Otherwise a
`RemoteBatchProperties` might be planned *twice* of which once is
unneeded for properties that are only needed after the mutating
operation.

If a mutating pattern does **not** invalidate the cached properties,
then property access after the mutating operator might already be
fetched earlier and need to be passed through to any parent operator.

## Categorization of the mutating patterns
The mutating patterns that do *not* invalidate the cached properties
are:
- `SetLabelPattern`
- `RemoveLabelPattern`
- `CreatePattern`
- `DeletePattern`

The mutating patterns that *might* invalidate cached properties are:
- `MERGE` - when one of the mutating patterns in the `onMatch`
invalidates cached properties
- `FOREACH` - when it contains a mutating pattern that invalidates
cached properties

All other mutating patterns affect properties and could therefore
*invalidate* the cached properties.