Fixes a double-upgrade bug ()
# Problem

* Given a database where the last tx in the tx log is an upgrade tx.
* When doing a backup and restoring that backup.
* Then we will start the server with the pre-upgrade kernel version,
leading to a double upgrade.
* This might succeed, or fail. It failed on a particular aura instance.
In any cause, a double upgrade in the TX log is a bug that needs fixing.

To start your review I recommend looking at the changes in
`RestoreDatabaseCommandIT`.

# Root cause

When doing the backup, we would create a new checkpoint with what should
be the latest kernel version. But that kernel version was the version of
the last entry in the tx log. In the case of an upgrade transaction, the
version of that log entry is the pre-upgrade-version, which is not the
version of the database if that log entry was successfully applied.

# Solution

Propagate the correct kernel version all the way:
 * As part of backup, we always do a checkpoint.
* `CheckpointerImpl.doCheckpoint` now includes the `kernelVersion` in
the `LatestCheckpointInfo`.
* This is read by `DefaultStoreSnapshotFactory.createStoreSnapshot` and
put into the `StoreSnapshot`.
* This is read by
`PrepareStoreCopyRequestHandler.SingleRequestHandler.sendUnrecoverableFilesAndConstructResponse`
and put into the `PrepareStoreCopyResponse`.
* This is encoded/decoded by `PrepareStoreCopyResponse.Encoder` /
`PrepareStoreCopyResponse.Decoder` if on the latest catch up protocol
`CATCHUP_PROTOCOL_V9`, which has been introduced in this commit. Older
catch up protocol versions will not read or write this field.
* `StoreFilesDownloader.downloadStoreFiles` reads either the new,
correct kernel version, or the old, incorrect kernel version through
`PrepareStoreCopyResponse.latestKernelVersion`, and puts it into
`StoreDownloadResult`.
* This is read by `RemoteStore.copy` and put into the
`CatchupSessionType.FullStoreCopy`.
 * This is read by `TxLogCatchUpSession.close`, which
   * creates a new TX log with the given version if there is none, and
* does ANOTHER checkpoint with the given version, to force recovery from
a given append index and transaction id.

# Notes and open questions

**Please answer these questions as part of your review.**

* `PrepareStoreCopyResponse.TransactionId` is broader than we need. We
previously used `.kernelVersion` (only for fallback on older protocol
versions after this commit). We still use `.id` and `.checksum`. The
other fields are unused and could be removed from the catch up protocol
if desired. This would also avoid that new code paths by accident use
`prepareStoreCopyResponse.lastCheckPointTx().kernelVersion` instead of
`prepareStoreCopyResponse.latestKernelVersion()`. I would consider this
follow-up cleanup to be handled by the clustering team, but please let
me know if you disagree.
* Do we need another version of the seed syncing protocol? As of now it
is using the now outdated `BaseCatchupClientDecodersV8` and
`BaseCatchupServerEncodersV8`, but I don't see any immediate issue with
that.
* I am unsure about the change in StoreFilesDownloader:100. I've changed
this to also use the "new" kernel version, i.e.
`prepareStoreCopyResponse.latestKernelVersion()`. It is used to
determine a log format in the absence of a
`prepareStoreCopyResponse.logFormat()`. That log version is then used to
create a new TX log if there is none in `TxLogCatchUpSession.close`. Is
this change correct?
*
`CatchupServerIT.shouldCommunicateErrorIfStoreIdDoesNotMatchRequest(CATCHUP_PROTOCOL_V9)`
is timing out and I need help debugging why. Neither the thread dump nor
the debug logs tell me anything useful.


Fixes KRNL-424