Block migrate with envelopes from downgrading
Migrate could pick the wrong log format if the log format setting had changed.
Since there is no logic for going from a higher to a lower format it blew up.
This change also starts makes sure to consider this an error in startup if it is ever wrongly configured again.