SPD property shard RPC serialize TxState directly to network channel ()
With this change we serialize/deserialize the filtered TxState directly
to/from the network channel instead of storing the serialized state in
an intermediate list.
That intermediate list effectively more than doubled the memory
consumption of the TX state/write transaction. This brings heap memory
consumption of write transactions closer to non-spd setups
It also creates a better separation between the RPC protocol and the TX
state serialization format.

Fixes KRNL-437
Related to KRNL-501