Set a static runtime for the inner queries for system commands ()
Uses slotted as it is available in both Community and Enterprise, and
handles all kinds of commands (non-admin show for example isn't covered
by pipelined).

Also sets cypher version for all inner queries:
* Read admin commands should use the Cypher version from the outer query
* Write admin commands and predicate checks should always use Cypher 25 for its inner query

And keep any pre-parser version option for the metadata queries.
These queries are admin commands themselves, so they don't need 
the `formatQuery` parts like we have for the inner queries of the commands.