Add temporal field support to Vector index single stage filter parameters ()
Initial set of types supported for fields did not include temporal.

This change adds support for temporal values, and for both point and
range queries.
- temporal values must have a zone (i.e. not local)
- values are converted to an Instant (UTC, more or less)
- values are stored as 12 bytes, encoding the long representing epoch
seconds, followed by the int representing nanos within that
- range queries work with a single query, as the encoding preserves
timeline order
- we do not store the zone value which produced the instant. Recall that
the Lucene document index is used to retrieve the node/relationship id
which is where the true values are found.

---------

Co-authored-by: Agent Smith <103414894+neo4jagentsmith[bot]@users.noreply.github.com>