Add AllPredicate and support in VectorIndex ()
For handling composite queries with gaps, i.e. if we index properties
p1, p2, ... but in the single-stage filter we only use a subset of the
properties we need a way to encode the "missing" properties. In theory,
we could just rely on `null` but that is problematic since a lot of
existing code interprets `null` as an impossible predicate rather than a
predicate matching everything. Instead, we introduce and explicit ALL
predicate to encode a predicate that matches everything, regardless if
the property exists or not. This PR only adds support for the new
predicate in the vector index, support in the other indexes will be
added in later PRs.