Lift restriction of indexes having a single type ()
This is like https://github.com/neo-technology/neo4j/pull/33948 but for
relationships. The only difference is that we don't actually have any
operators with multiple types yet. In an effort to reduce the size of my
PRs, I will add the relationship variant(s) of the vector search
operator in a separate PR.