Rewrite PropertiesUsingCachedProperties to non-cached ()
During expression variable allocation sometimes cached properties are
rewritten to non-cached properties. If the expression also contained a
`PropertiesUsingCachedProperties` function, its set of cached properties
would get rewritten to an incompatible `Property` type. Later when that
set was iterated it contained objects of an unexpected type and a class
cast exception was thrown.

This fix will rewrite the `PropertiesUsingCachedProperties` back to the
non-cached function call.