BlockBatchInput no longer gives relationship IDs to importer ()
It was a remnant from the legacy way that SPD import worked.
Relationship IDs must be generated by the importer itself. Using the
relationship IDs that were given from the outside could cause huge
entity properties to be written for the wrong relationship ID, resulting
in inability to read them later.

cl Fixes an issue when running e.g. copy or migrate where the source and
target database is block format, where relationships with a large amount
of properties might lose their properties.