Don't pick MVCC components based on config ()
There are some places that check for the db_format from the config to
make decisions about what components to use. This is incorrect, as there
might already be store files, after a restart or a restore. I found this
while investigating why multiversion wasn't being picked up when
restoring/store copying a multiversion db.

* `Stores` should use the format from an existing store, or else the
config. This was recently fixed, but there was one case that wasn't
covered, so I fixed that and added a test.
* `CursorContextFactory` can't be picked at construction time (in the
various `ContextFactory`s. Similar to `StorageEngineFactory`, we miss
the information we need, i.e. which store will be used, to make a
correct decision. Therefore I made it lazy, so it's only constructed
when we know which SEF to use, i.e. at Database.init().
* `IdContextFactory` and a few other components is the same as above. 

I've piggy backed on the `StorageEngineFactory` to know if the database
is MVCC or not. Because the `StorageEngineFactory` initialisation is
already lazy, and happens after cluster has synced the stores, we know
that the store on disk is valid and will be used. Also the SEF
initialisation reads the local store and picks the correct SEF, so if
that one says MVCC the database is truly MVCC. If there are no previous
files, the SEF initialisation (`StorageEngineFactorySupplier`) gets the
format from either user input or config.

Most components were easy to make lazy, but the `CursorContextFactory`
was trickier. It's used in cluster components (RemoteStore, StoreFiles)
before Database.init() and therefore I needed to create an "offline"
version of the `CursorContextFactory`, meaning that it doesn't have a
`VersionContext`, which should be fine at startup in those components.
But once `CursorContextFactory` has been created, that one will be used
in those components too.