Don't sort twice if the show command have an ORDER BY in its YIELD ()
If we don't sort in the YIELD we need to give back the rows in the
default sort order, but if we sort in the YIELD that order is irrelevant
so no need for the extra default order sorting.

An `ORDER BY` on the return shouldn't affect the default sorting as the
rows returned from one show command need to have either an explicit
sorting or the default sorting when incoming to other clauses, for
example `SHOW FUNCTIONS YIELD name AS func SHOW PROCEDURES YIELD name AS
proc RETURN func, proc ORDER BY func, proc` (currently allowed behind
feature flag) or `SHOW INDEXES YIELD entityType, createStatement WITH
createStatement WHERE entitiyType = 'NODE' RETURN * ORDER BY
createStatement DESC` (currently ongoing DWYW project).