Save database name and uuid early ()
Store the name and uuid of the database about to be dropped when
checking that it exists.

The database could in some cases be dropped between the check that it
exists and actually dropping it. But it was at the point of the drop
that the name and uuid was stored in the context to later be used in the
wait part of the plan.