feat(VectorSearch) - Throw GQL code 22G03 with cause 22N27 when using vector search with a path variable ()
part of PLAN-2900

Providing a bounding variable of a type that is always wrong (not a NODE
or a VARIABLE) gives GQL error code `22G03` with cause `22N27`.

Providing a bounding variable that is of a NODE or VARIABLE but the
wrong one with respect to the vectorIndex gives GQL error code `22G03`
with cause `22N01`.