KRNL-484 - Fix flaky ThreadSanitizerTest for sequential access ()
**Issue:** https://linear.app/neo4j/issue/KRNL-484/nightly-test-failure

This test was set up such that it waits for `typeADoThing` (from within
`doThing`) to count down on the separate thread before running `doThing`
in my main thread. The problem was that I hadn't guaranteed that the
other thread had exited `typeADoThing` - it was possible that execution
was still at `latch.await()` in that method while the test thread
proceeds onto running the same method in-thread.

The solution, therefore, is to move the `countdown` in this test to
after the sanitised method, whilst still being in the `doThing`
Runnable. That way, I can guarantee no parallel access in this test.