Tests for variable checking of path variables in scalar subqueries ()
This adds some test to the variable checker, which is a strong
indication that SURF-137 will fix simply by switching to completely to
the new way of variable checking (and clean out the over code).

Hence, I propose to add these three tests as a way of putting a stake
into the ground.

Part of SURF-137