add round trip check to PrettifierIT ()
This PR adds a roundtrip check to the prettifier. That is

cypher -> parse -> AST -> prettify -> parse -> AST'

should work such that AST = AST'.

Canonicalization of syntax done by the parser or prettifier prevents a
successful round trip for certain syntax elements. The test accounts for
these in `AbstractPrettifierTest.rewriteASTDifferences`. A method that
likely needs to be extended when new syntax elements with
canonicalizations are added to Cypher.