Simplify deleteAfterCopy in Dumper ()
It was possible to pass different values in the option to
`Dumper.visitPath`, but in practice it was always used as either all
true, or all false.

Instead we assign this state to the dumper, and simply unlink the file
directly after it has been processed, if that was the requested action.
There is no need to keep this state in the visitor/operations.

---------

Co-authored-by: MishaDemianenko <12656787+mishademianenko@users.noreply.github.com>