Run ScopeSurveyor in production pipeline ()
Integrates the ScopeSurveyor into the transformer pipeline. There is
currently no consumer and the scopes are only kept up to date until
after the ParsePipelineTransformers are done. This is to limit overhead.
Any future consumer will have to make sure that it is receiving up to
date scopes by using the `UpToDateScopes` condition.