Fix for flaky spd batch test ()
The problem was that during txState serialization to get the physical
commands we used the transaction tracker in parallel from different
threads, causing the heap leaks. I changed to using the empty memory
tracker during visit, since the memory tracker from the request is meant
to be the property shard one. So to use the correct tracker we would
have to change the request to end in both a graph shard tracker and
property shard one. If we did that we would have to use the
`shardResponseMemoryTracker` in the `ExtendedSpdClient` for it to be
thread safe
Fixes: KRNL-611, KRNL-610, KRNL-609