Reuse deleted non-visible data layer pages in MVCC ()
In a multi-versioned multi-root GBPTree, we can't release the data layer
pages on deletion time, since we need to keep the visibility until
oldestVisibleTransaction is past. Currently we just leave it behind.
With this PR we will release them with a condition. There's already the
condition that the generation in which it was released has become
stable. Now we also wait until the version in which it was released has
been seen by all transactions. Note that this extra condition is only
used in multi-versioned trees, when deleting the rootKey. For all other
cases, the behaviour should not be changed.

This in combination with the
https://github.com/neo-technology/neo4j/pull/32928 that cleans up the
rootKey itself, the MultiRootGBPTree should not leave garbage behind for
deleted rootKeys.