Default value for includesChangesFromThisTransaction is true if not mutliversioned cursor ()
Default mode for cursors is to include changes from the current
transaction. Not multiversion cursors also should have it as "include"
despite they do not see it to simplify preserving of the existing logic.

Fix `MuninnReadPageCursor` to follow this so `BlockNodeCursor` do not
reload page cursor on every iteration.

Fixes KRNL-699