Support relationship vector searches in slotted (and kind of legacy) ()
Add spec-test + implementation in slotted and legacy. Since vectors only
work on block format and we run legacy tests in community where we only
have record format, no tests aren't running with the legacy runtime.
However, the implementations almost share all the code so I think it is
fine but if you disagree we can a variant that runs legacy-runtime in
enterprise?

Pipelined/Parallel version coming soon in a separate PR.