Index updates no longer cause OOM during apply. ()
Fixes KRNL-433

## The problem

When a transaction's commands barely fit in memory, but we successfully
append to log, and when we then generate many index commands during
appy-to-store, we crash with an OOM that causes a database panic since
the TX was already appended to the log.

### Testing problem

Since we don't track memory of index updates, the change was hard to
test. This PR does not aim to introduce memroy tracking there. The added
test relies on configuring a lower heap limit for the JVM for that
single test and would prior to the changes OOM.

## Solution

### Release memory earlier in IndexUpdatesWorkSync
(64ac39d90eea3595ec00b5cf09dd0eefcffc7feb)

One part of the solution is to release memory of index updates that have
already been processed as soon as possible.
This requires some changes to provide an `Iterator` instead of an
`Iterable` to the `IndexingService` to make sure no one attempts to
iterate miltiple times through the updates.

### Use a lazy version of `ValueIndexEntryUpdate`

Instead of reading all the `Value`s already in
`BlockTransactionApplier`, we read them later (using a `Supplier`) when
actually applying the index updates. This avoids holding onto double
memory for a long time (double because we still hold onto the buffer
from which we read the Value, and the Value).
Together with releasing the index updates earlier, this solves the
problem.

The lazy values are cached once they are computed for the first time.
This is to avoid regressions, because otherwise we would read the value
multiple times, e.g. if a value is part of many indexes.

Please read the commit descriptions for more detail.

## Tested alternatives

Don't cache the values from the Suppliers. This would cause regressions
though, when we have to read values that are part of many indexes
multiple times. It also did not work with the importer, for unclear
reasons.