EphemeralFileSystemAbstraction also synchronizes deletions ()
The test com.neo4j.HalfAppliedConstraintRecoveryIT was flaky, and it
turns out that it failed when making a snapshot of its
EphemeralFileSystemAbstraction in the test. It failed because some file
got deleted while doing the snapshot. There are a certain set of files
that can come and go, in this case it was a temporary "scan data" file
used when building an index.
Most methods that can create files are synchronized, however the methods
that deletes files weren't, allowing this race to happen.

Fixes KRNL-696, KRNL-697