Fix Bedrock integration tests ()
The region used was set as an env, but this means all Bedrock tests must
work in the same region. AWS has models that are not available in all
regions and therefore it is very difficult to test Titan and Nova in the
same region for text completion. As the region used is not a secret, I
have hardcoded it instead to work (even with the original env way,
updating would've been a change to code, so it does not make it harder
to fix in case of a model being dropped from a region)