SystemGraphComponent for fleet management module ()
This pull request implements the `SystemGraphComponent` interface for
the fleet manager module, and wires it into the community and enterprise
edition factory modules.

Future versions of the schema will introduce an implementation of
`KnownFleetManagementComponentVersion` and override a method of the
following signature:
`public void upgradeTopologyGraph(Transaction tx, int fromVersion)
throws Exception`

This method will perform any data migration necessary from the previous
schema version. This original revision ships
`FleetManagementComponentVersion_0_202511` which only sets the version
(to 0). If the version is uninitialized, it will be upgraded to v0
through this flow.