Provide type information with the load input ()
Future format updates, will have to know about if the input comes from
stdin, or if comes from a path, since we may have to treat these two
cases somewhat differently.

This commit reuses the idea of `DumpInput` from LoadDumpExecutor, but
reimplements it as two different record types sharing an interface.

This allows us to clean up the code a bit and remove some duplication.