COM-1009: Allocates constraint schema rule IDs earlier ()
When creating an index, its schema rule ID is assigned to it right away
(e.g. in Operations#indexDoCreate), and therefore those IDs already are
in the TxState at commit.

For some reason this wasn't the case for constraints - instead they were
assigned IDs when creating commands. This meant that in SPD when it
serialized the TxState over to the property shards it didn't carry with
it the constraint schema rule IDs. This forced the property shards to
allocate their own, and they wouldn't necessarily be the same - what
with ID reuse and all.

This might have been fine to do if not for uniqueness constraints which
links to an "owned index ID", and THAT id was carried over to property
shards. This resulted in property shards having their own set of
constraint IDs, but sharing links to the graph shard's index IDs, and it
could result in a constraint having the same ID as the index it linked
to.