Forward/Reverse recovery until position
Introduces a new recovery predicate that can do recovery to a specific position.
Note that nothing past this position is allowed to have been applied or partially
applied to the stores since backwards recovery is only done from the position.
This is introduced to be able to have a live log while recovery happens.

This part does not update the logtail reading and isRecoveryRequired checks, that will come separately.