Make seed from URI abortable ()
Fixes CLUSTER-460

For a long time the idea was that we need to inject the aborter to every
piece of code which is involved in the seeding. This seemed like a bit
impossible.
In a recent [spike](https://github.com/neo-technology/neo4j/pull/34265)
a similar issue (something needs to be aborted which was not designed
for it) was solved with the help of the `JobScheduler` and that one can
"cancel" those jobs. Nothing stops us to do the same here.

No extra test was written here, the already existing
`CatchupSeedProviderIT.shouldAbortSeedingWhenTargetDatabaseStopped`
demonstrates that this is working indeed.

Tests for the `AbortableRunner` will come with the other PR.