Changing the eviction algorithm of TransactionMetadataCache
It behaves like a cache with a least-recent-added eviction policy now.
This simple caching mechanism replaces Caffeine-based LFU Cache. The
behaviour of Caffeine was unsuitable for transaction metadata caching.

Caffeine cache is split into two - a small cache holding recently added
records and the main cache. When the main cache is full, a record is
promoted from the recently added cache into the main one only if it has
number of usages higher than the record it would replace. This is
problematic for this use case, because transaction metadata is mostly
written to the cache and very rarely looked up. So new records usually
don't make it to the main cache. This contrasts the usage of this cache
where recently added records are the most useful.

This change makes SPD write SOAKs 5x faster.
Another benefit is that the cache takes significantly less memory now.