Use real oldestTransactionId in cluster commit process ()
We have to use the real value, and not 0. Since it's already at commit
phase, nothing will be read, so we think it's safe to use the real
value. And when writing a new value, we don't want to save history that
no one will read. OldestTransactionId being 0 basically means save
everything. This is both unnecessary, and can cause inconsistencies in
the history (bad magic for example).