Vector index SSF temporal cypher compatibility ()
Temporal parameters have already been added to Vector Single Stage
Filters as types for which the properties of nodes and relationships
within vector indexes can be filtered for. The simple rule for the
initial implementation was to order temporals based on the instant which
they represent. Temporals with the same instant were considered equal.

In Cypher, the rules are different. It has been decided that we should
conform to the Cyper rules.


https://docs.google.com/document/d/1zTlmGMzP0e82zXUHlKd4V6rfYogSvbMtufP75ooSLL4/edit?tab=t.0#heading=h.ceo00btrllhd

Where temporals have the same instant, the order is further
distinguished:
1. By the timezone offset, if the temporal value has such a thing (times
and datetimes do)
2. If values are still equal, by the timezone id, if the temporal value
has such a thing (datetimes do). Timezone ids exist in 2 forms, as a
simple offset id, or as a region id. Offset ids order before region ids,
and within offset ids and region ids, the order is by the alphanumeric
string value of the id.

To implement this, we add (when the values exist) 2 new Lucene index
fields per value,
1. for the offset (as an integer, in seconds)
2. for the zone id, either `Offset::<zone-id>` or `Region::<zone-id>`

Applying queries to the vector index for exact queries, we need a
conjunction of the (instant, offset, id) values in the query.
Range queries are rather more involved, being one of:
- A range of instants where all values with that instant value match
- A range of timezone offsets for the endpoint instants (one semi-open
range at each end)
- A range of timezone ids for the endpoint offset within the endpoint
instant (one semi-open range at each end)

This closes IND-107