Support single-stage filtering of Duration ()
This does the most straight-forward implementation by storing all the
four fields of a duration value, `months`, `days`, `seconds` and `nanos`
in the document and then queries for exactly these. We don't support
comparing durations so range queries will not match any documents.

We could optimize this a bit by only storing non-zero fields, I think in
most cases the duration will only have a second field, and then adapt
the query to search for only non-zero fields + the non-existence of the
zero fields.

Fixes IND-143