Don't print predicate string, nor seed, in test name ()
The predicate can contain weird characters that fails xml parsing on
team-city. Also remove the seed from the test name to be nicer to the
flaky test detection, now we shouldn't get one card per seed. Instead,
we just println the values so that we can find them in the log in case a
failure would happen.