feat(VectorSearch): Plan type filters if vectorindex has a different type ()
Other relationship leaf planners use
`RelationshipLeafPlanner.planHiddenSelectionAndRelationshipLeafPlan`
However vector search is different. The key differences are that 
1. other planners use the relationship type from the
patternRelationship. However, with vectorSearch we rely on the index for
the type description, otherwise the search will never get planned.
Therefore for vectorsearch, the different types should be verified using
a filter later on.
2. other planners plan hidden selections on bound nodes. Vector search
currently only allows one binding variable per clause, which for
relationships has to be to the relationship variable. So the bound nodes
filter should not be required ideally.

---------

Co-authored-by: Arne Fischereit <arne.fischereit@neo4j.com>